ThanhGPT_Backend_FastAPI/requirements.txt
fastapi
uvicorn
openai
pytesseract
pillow